LayoutTests/svg/css/parse-calc-length-expected.txt

CONSOLE MESSAGE: Error: Invalid value for <rect> attribute height="calc(500px - 400px)"
CONSOLE MESSAGE: Error: Invalid value for <rect> attribute rx="calc(500px - 400px)"
CONSOLE MESSAGE: Error: Invalid value for <rect> attribute ry="calc(500px - 400px)"
CONSOLE MESSAGE: Error: Invalid value for <rect> attribute width="calc(500px - 400px)"
CONSOLE MESSAGE: Error: Invalid value for <rect> attribute x="calc(500px - 400px)"
CONSOLE MESSAGE: Error: Invalid value for <rect> attribute y="calc(500px - 400px)"
CONSOLE MESSAGE: Error: Invalid value for <rect> attribute height="calc(40px + 60px)"
CONSOLE MESSAGE: Error: Invalid value for <rect> attribute rx="calc(40px + 60px)"
CONSOLE MESSAGE: Error: Invalid value for <rect> attribute ry="calc(40px + 60px)"
CONSOLE MESSAGE: Error: Invalid value for <rect> attribute width="calc(40px + 60px)"
CONSOLE MESSAGE: Error: Invalid value for <rect> attribute x="calc(40px + 60px)"
CONSOLE MESSAGE: Error: Invalid value for <rect> attribute y="calc(40px + 60px)"
CONSOLE MESSAGE: Error: Invalid value for <rect> attribute height="calc(0% + 100px)"
CONSOLE MESSAGE: Error: Invalid value for <rect> attribute rx="calc(0% + 100px)"
CONSOLE MESSAGE: Error: Invalid value for <rect> attribute ry="calc(0% + 100px)"
CONSOLE MESSAGE: Error: Invalid value for <rect> attribute width="calc(0% + 100px)"
CONSOLE MESSAGE: Error: Invalid value for <rect> attribute x="calc(0% + 100px)"
CONSOLE MESSAGE: Error: Invalid value for <rect> attribute y="calc(0% + 100px)"
CONSOLE MESSAGE: Error: Invalid value for <rect> attribute height="calc(100% - 100px)"
CONSOLE MESSAGE: Error: Invalid value for <rect> attribute rx="calc(100% - 100px)"
CONSOLE MESSAGE: Error: Invalid value for <rect> attribute ry="calc(100% - 100px)"
CONSOLE MESSAGE: Error: Invalid value for <rect> attribute width="calc(100% - 100px)"
CONSOLE MESSAGE: Error: Invalid value for <rect> attribute x="calc(100% - 100px)"
CONSOLE MESSAGE: Error: Invalid value for <rect> attribute y="calc(100% - 100px)"
Test that length values for presentation attribute accept calc.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS computedStyle("baseline-shift", "calc(500px - 400px)") is "100px"
PASS computedStyle("cx", "calc(500px - 400px)") is "100px"
PASS computedStyle("cy", "calc(500px - 400px)") is "100px"
PASS computedStyle("height", "calc(500px - 400px)") is "100px"
PASS computedStyle("kerning", "calc(500px - 400px)") is "100px"
PASS computedStyle("r", "calc(500px - 400px)") is "100px"
PASS computedStyle("rx", "calc(500px - 400px)") is "100px"
PASS computedStyle("ry", "calc(500px - 400px)") is "100px"
PASS computedStyle("stroke-width", "calc(500px - 400px)") is "100px"
PASS computedStyle("stroke-dashoffset", "calc(500px - 400px)") is "100px"
PASS computedStyle("width", "calc(500px - 400px)") is "100px"
PASS computedStyle("x", "calc(500px - 400px)") is "100px"
PASS computedStyle("y", "calc(500px - 400px)") is "100px"
PASS computedStyle("baseline-shift", "calc(40px + 60px)") is "100px"
PASS computedStyle("cx", "calc(40px + 60px)") is "100px"
PASS computedStyle("cy", "calc(40px + 60px)") is "100px"
PASS computedStyle("height", "calc(40px + 60px)") is "100px"
PASS computedStyle("kerning", "calc(40px + 60px)") is "100px"
PASS computedStyle("r", "calc(40px + 60px)") is "100px"
PASS computedStyle("rx", "calc(40px + 60px)") is "100px"
PASS computedStyle("ry", "calc(40px + 60px)") is "100px"
PASS computedStyle("stroke-width", "calc(40px + 60px)") is "100px"
PASS computedStyle("stroke-dashoffset", "calc(40px + 60px)") is "100px"
PASS computedStyle("width", "calc(40px + 60px)") is "100px"
PASS computedStyle("x", "calc(40px + 60px)") is "100px"
PASS computedStyle("y", "calc(40px + 60px)") is "100px"
PASS computedStyle("baseline-shift", "calc(0% + 100px)") is "0px"
PASS computedStyle("cx", "calc(0% + 100px)") is "calc(0% + 100px)"
PASS computedStyle("cy", "calc(0% + 100px)") is "calc(0% + 100px)"
PASS computedStyle("height", "calc(0% + 100px)") is "calc(0% + 100px)"
PASS computedStyle("kerning", "calc(0% + 100px)") is "0"
PASS computedStyle("r", "calc(0% + 100px)") is "calc(0% + 100px)"
PASS computedStyle("rx", "calc(0% + 100px)") is "calc(0% + 100px)"
PASS computedStyle("ry", "calc(0% + 100px)") is "calc(0% + 100px)"
PASS computedStyle("stroke-width", "calc(0% + 100px)") is "calc(0% + 100px)"
PASS computedStyle("stroke-dashoffset", "calc(0% + 100px)") is "calc(0% + 100px)"
PASS computedStyle("width", "calc(0% + 100px)") is "calc(0% + 100px)"
PASS computedStyle("x", "calc(0% + 100px)") is "calc(0% + 100px)"
PASS computedStyle("y", "calc(0% + 100px)") is "calc(0% + 100px)"
PASS computedStyle("baseline-shift", "calc(100% - 100px)") is "0px"
PASS computedStyle("cx", "calc(100% - 100px)") is "calc(100% - 100px)"
PASS computedStyle("cy", "calc(100% - 100px)") is "calc(100% - 100px)"
PASS computedStyle("height", "calc(100% - 100px)") is "calc(100% - 100px)"
PASS computedStyle("kerning", "calc(100% - 100px)") is "0"
PASS computedStyle("r", "calc(100% - 100px)") is "calc(100% - 100px)"
PASS computedStyle("rx", "calc(100% - 100px)") is "calc(100% - 100px)"
PASS computedStyle("ry", "calc(100% - 100px)") is "calc(100% - 100px)"
PASS computedStyle("stroke-width", "calc(100% - 100px)") is "calc(100% - 100px)"
PASS computedStyle("stroke-dashoffset", "calc(100% - 100px)") is "calc(100% - 100px)"
PASS computedStyle("width", "calc(100% - 100px)") is "calc(100% - 100px)"
PASS computedStyle("x", "calc(100% - 100px)") is "calc(100% - 100px)"
PASS computedStyle("y", "calc(100% - 100px)") is "calc(100% - 100px)"
PASS computedStyle("stroke-miterlimit", "calc(500 - 400)") is "100"
PASS computedStyle("flood-opacity", "calc(0.75 - 0.25)") is "0.5"
PASS successfullyParsed is true

TEST COMPLETE
